Settled on elasticsearch UI library for building the kibana components
So kinda in a wierd place.  Would prefer to use material-ui or some other react library because Elasticsearch has a history of just ripping stuff up so it will be stable.  But we are on react 16.2, which mean we would need to use a beta pre 1.0 of material-ui (which was 5 months ago, they are on 3.1 now) or i can just roll my own, which i'm really not wanting to do.